fastapi/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.13.1
anyio==3.7.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Cython==3.0.2
decorator==5.1.1
distro==1.8.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.0
fonttools==4.38.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.18.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
greenlet==3.0.3
grpcio==1.54.2
grpcio-status==1.54.2
gTTS==2.3.2
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
idna==3.4
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.4
mysql-connector-python==8.2.0
numpy==1.24.1
oauth2==1.9.0.post1
openai==1.2.3
orjson==3.9.0
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.27.0
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.9.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wavio==0.0.7
wcwidth==0.2.6
websockets==11.0.3
yarl==1.8.2
